+++ lib/tab_switcher.dart
@@ -1,3 +1,3 @@
-export 'controller.dart';
-export 'impl.dart';
-export 'widgets/theme.dart';
+export 'utils/controller.dart';
+export 'utils/impl.dart';
+export 'utils/theme.dart';
