+++ lib/widgets/tab_grid.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:scroll_shadow_container/scroll_shadow_container.dart';
 
-import '../controller.dart';
+import '../utils/controller.dart';
+import '../utils/theme.dart';
 import 'animated_grid.dart';
 import 'minimized_tab.dart';
-import 'theme.dart';
 
 /// Displays grid of minimized tabs
 class TabSwitcherTabGrid extends StatefulWidget {
